Fixed build failures for DLL

+++ ApiTest/descriptorheap.cpp
@@ -50,7 +50,7 @@ bool Test18(_In_ ID3D12Device* device)
             HRESULT hr = device->CreateDescriptorHeap(&desc, IID_PPV_ARGS(&existingHeap));
             if (FAILED(hr))
             {
-                printf("ERROR: Failed to create test descriptor heap (hr: 0x%08X)\n", static_cast<unsigned long>(hr));
+                printf("ERROR: Failed to create test descriptor heap (hr: 0x%08X)\n", static_cast<unsigned int>(hr));
                 return false;
             }
 
@@ -195,7 +195,7 @@ bool Test19(_In_ ID3D12Device* device)
     HRESULT hr = device->CreateDescriptorHeap(&desc, IID_PPV_ARGS(&existingHeap));
     if (FAILED(hr))
     {
-        printf("ERROR: Failed to create test descriptor heap (hr: 0x%08X)\n", static_cast<unsigned long>(hr));
+        printf("ERROR: Failed to create test descriptor heap (hr: 0x%08X)\n", static_cast<unsigned int>(hr));
         return false;
     }
